+++ packages/cap-config-writer/src/cap-writer/index.ts
@@ -1,3 +1,4 @@
 export { updateAppPackageJson, updateRootPackageJson } from './package-json';
 export { updateTsConfig, updateStaticLocationsInApplicationYaml } from './tsconfig-and-yaml';
 export { updatePomXml } from './pom-xml';
+export { getAppLaunchText } from './utils';