+++ packages/preview-middleware-client/src/cpe/changes/validator.ts
@@ -10,7 +10,7 @@ import type UI5Element from 'sap/ui/core/Element';
  */
 export function validateBindingModel(modifiedControl: UI5Element, value: string): void {
     const bindingValue = value.replace(/[{}]/gi, '').trim();
-    const bindingParts = bindingValue.split('>').filter((el) => el != '');
+    const bindingParts = bindingValue.split('>').filter((el) => el !== '');
 
     if (!bindingParts.length) {
         throw new SyntaxError('Invalid binding string.');
@@ -22,7 +22,8 @@ export function validateBindingModel(modifiedControl: UI5Element, value: string)
             const resourceBundle = (modifiedControl.getModel('i18n') as ResourceModel).getResourceBundle() as ResourceBundle;
             if (!resourceBundle.getText(resourceKey, undefined, true)) {
                 throw new SyntaxError(
-                    'Invalid key in the binding string. Supported value pattern is {i18n>YOUR_KEY}. Check if the key already exists in i18n.properties. If not, add the key in the i18n.properties file and reload the editor for the new key to take effect.'
+                    'Invalid key in the binding string. Supported value pattern is {i18n>YOUR_KEY}. Check if the key already exists in i18n.properties.' +
+                        'If not, add the key in the i18n.properties file and reload the editor for the new key to take effect.'
                 );
             }
         } else {