+++ packages/fiori-generator-shared/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @sap-ux/fiori-generator-shared
 
+## 0.1.1
+
+### Patch Changes
+
+-   Updated dependencies [9ea58ad4]
+    -   @sap-ux/project-access@1.22.4
+
+## 0.1.0
+
+### Minor Changes
+
+-   3b795703: Adding generate readme file functionalities
+
+## 0.0.17
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.22.3
+
+## 0.0.16
+
+### Patch Changes
+
+-   Updated dependencies [399d2ad8]
+    -   @sap-ux/project-access@1.22.2
+
+## 0.0.15
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.22.1
+
 ## 0.0.14
 
 ### Patch Changes