+++ packages/ui5-application-writer/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sap-ux/ui5-application-writer
 
+## 0.26.18
+
+### Patch Changes
+
+-   Updated dependencies [1a1baeb0]
+    -   @sap-ux/ui5-config@0.22.10
+
+## 0.26.17
+
+### Patch Changes
+
+-   Updated dependencies [399d2ad8]
+    -   @sap-ux/ui5-config@0.22.9
+
+## 0.26.16
+
+### Patch Changes
+
+-   Updated dependencies [a140cf8b]
+    -   @sap-ux/ui5-config@0.22.8
+
 ## 0.26.15
 
 ### Patch Changes